pyproject.toml
[project]
name = "ollim-bot"
version = "0.1.0"
description = "ADHD-friendly Discord bot with proactive reminders, powered by Claude"
requires-python = ">=3.11"
license = "GPL-3.0-only"
readme = "README.md"
authors = [{ name = "Julius" }]
keywords = ["discord", "adhd", "claude", "productivity", "bot", "reminders"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Framework :: AsyncIO",
    "Topic :: Communications :: Chat",
]
dependencies = [
    "discord.py>=2.6.4",
    "claude-agent-sdk>=0.1.39",
    "google-api-python-client>=2.190.0",
    "google-auth-oauthlib>=1.2.4",
    "apscheduler>=3.11.2",
    "python-dotenv>=1.2.1",
    "google-auth>=2.48.0",
    "pyyaml>=6.0.3",
    "aiohttp>=3.13.3",
    "jsonschema>=4.26.0",
]

[project.urls]
Repository = "https://github.com/julius/ollim-bot"
Issues = "https://github.com/julius/ollim-bot/issues"

[project.scripts]
ollim-bot = "ollim_bot.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.11.0",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM", "RUF"]
ignore = ["SIM102", "SIM108", "E501"]

[tool.ruff.lint.isort]
known-first-party = ["ollim_bot"]
